requirements.txt
aiohttp
aiomysql
aiosignal
annotated-types
anyio
astunparse
async-timeout
attrs
cryptography
Cython
dbutils
fastapi
h11
httpcore
httptools
httpx
jedi
json5
libcst
loguru
markdown-it-py
matplotlib
mdurl
multidict
numpy
omegaconf
openai<3.0,>=2.0
openpyxl
orjson>=3.0
pandas<3.0,>=2.0
parso
pillow
pymysql
pyparsing
python-dateutil
pyyaml
redis
requests
rich
sqlalchemy
starlette
typing-extensions
typing-inspection
tzdata
urllib3
uvicorn
uvloop
websockets
yarl
